InputCommon: Move SDL Device class into its own cpp/h files.
